Implement Frontend and Endpoints. This commit implements the Addon frontend which can be accessed in the Ayon Server > Settings > Shotgrid Sync page.
The frontend offers two actions, import and sync.
For Importing, it will, via the Shotgrid REST API implemented in the
callback for the `get-importable-projects` endpoint, fetch the projects
that are not a template, and will allow the person to choose from a
drodown the project to import, by clicking the "Import Project" button.

For Syncronizing, a call to the Ayon server GraphQl API is performed,
and we only display the projects that have the `shotgridId` attribute
filled, once the person picks a show and click the button
"Sync Shotgrid Project".

In both instances, upong clicking the button, will dispacth an event of
the topic `shotgrid.event` which has to be handled by a `processor`
respectively, which are not included in this commit, these are catched
by the `create-project` and `sync-from-shotgrid` endpoints.

The CSS and Javascript have been split in separate files.