`services.processor` fix event topic and payload
We are now handling all related Shotgrid events by dispatching them with
the shotgrid.event topic and providing all the information on the
payload, this commit addresses that as well as passing the payload dict
by unpacking it for use within the handlers, this means that handlers
have to accept a `**kwargs` on top of their required arguments.

A way to recreate cleanly the docker image has been added to the Dockerfile
and the `pyproject.toml` has been fixed to pull the lates `ayon-python-api`.