`create_project` and `sync_from_shotgrid` handlers This commit implements two handlers, which are interconected.
Firstly there's the `create_project`, triggered when a `shotgrid.event`
with the `action` payload of `create-project` is dispatched, which
creates a Shotgrid project in Ayon, it will also replicate the Shotgrid
entities and tasks in Ayon, by creating folders and tasks that match
these in the project anatomy.
It also sets up the Shotgrid project creating custom fields (all
prepended by `sg_ayon_`) in Shotgrid and trigger a full Sync once
everything is in place.

Secondly, there's the `sync_from_shotgrid`, triggered when a
`shotgrid.event` with the `action` payload of `sync-from-shotgrid`
is dispatched, which will perform a full import of a Shotgrid project
into Ayon, it will fail if there are missing attributes in Shotgrid,
which should be present unless tampered manually, it won't go through
otherwise.

In both instances the actual syncronization is performed by the class
`SyncFromShotgrid` found in the `lib/sync_from_shotgrid.py` module, this
class performs the checks, and traverses the Shotgrid project in the look
for entities to be created or updated in Ayon.

A set of `constants` and `utils` are defined in the `lib` directory,
the former contains Shotgrid field codes, Ayon attributes names and
 mappings between the two, and the latter includes helper functions to
query and marshall data from Shotgrid, as well as creating entities in
Ayon.