`update_from_shotgrid` handler on Shotgrid Events This commit implements the `update_from_shotgrid` handler, triggered when a `shotgrid.event` with the `action` payload of `shotgrid-event` is dispatched, which will attempt to either `create`, `update` or `remove` an Ayon entity, that corresponds to the Shotgrid entity that the event is acting on.
Shotgrid has the ability to "retire" (move to trash) entities, and then
bring them back, in those instances we treat it as if we are creating a
new entity, since we don't have that feature in Ayon.

As of this implementation, the `update_entity` only cares for the entity
name since it's all we really track at the moment, and ensures the `id`
mapping between Ayon and Shotgrid is still valid.