`client.tray` Improve the Tray actions.
Actions will change their text depending on if we are logged in or not,
and replaced `Shotgrid Manager` with the two new urls where the person
can manage shotgrid or trigger syncronization.

Read the Shotgrid URL from the actual module settings, and check that
the person can login with the provided username and password.