Implement `pyblish` plugins This commit updates and ensures the existing plugins work with Ayon. They have been cleaned up and streamlined where possible.
Full publish is not possible as of this writing since some bits of the
core app are not working yet, but if any error where to arise, it should
be fairly trivial to get them to work.

Ideally in the future we would collect the Shotgrid Version in the collect
entities step.