`server.frontend` - Improved UX for the Addon frontend This commit replaces the current UI for the "Shotgrid Sync" page, but it also removes logic previously done behind addon endpoints; this improves the code by reducing the redundant logic and the latency between the actions and the UI changes.
The interface has been changed to also be a table of all projects in both
AYON and Shotgrid with a column that tells the user if they exists in AYON
and/or Shotgrid, this is now all done in Javascript, which makes the endpoint
`"get-importable-projects"` redundant.

Another column "Syncronize" whith a button that will trigger
the dispatch of the "shotgrid.event" that the processor picks up, this makes
endpoint `"sync-from-shotgrid/{project_name}"` redundant.